requirements.txt

numpy==1.26.4
Pillow==9.2.0
tensorflow==2.11.0
opencv-python==4.8.1

python-dotenv
python-multipart

fastapi==0.110.0
uvicorn==0.28.0
